+++ executor/memtable_reader_test.go
@@ -16,17 +16,12 @@ package executor_test
 
 import (
 	"context"
-	"crypto/tls"
-	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"net"
-	"net/http"
 	"net/http/httptest"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync/atomic"
 	"time"
@@ -38,12 +33,8 @@ import (
 	"github.com/pingcap/kvproto/pkg/diagnosticspb"
 	"github.com/pingcap/sysutil"
 	"github.com/pingcap/tidb/domain"
-	"github.com/pingcap/tidb/executor"
 	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/planner/core"
 	"github.com/pingcap/tidb/session"
-	"github.com/pingcap/tidb/store/helper"
 	"github.com/pingcap/tidb/util/pdapi"
 	"github.com/pingcap/tidb/util/testkit"
 	pmodel "github.com/prometheus/common/model"
@@ -963,665 +954,4 @@ func (s *testMemTableReaderSuite) TestTiDBClusterLogError(c *C) {
 	c.Assert(err, NotNil)
 	c.Assert(err.Error(), Equals, "denied to scan full logs (use `SELECT * FROM cluster_log WHERE message LIKE '%'` explicitly if intentionally)")
 	c.Assert(rs.Close(), IsNil)
-}
-
-type mockStoreWithMultiPD struct {
-	helper.Storage
-	hosts []string
-}
-
-var hotRegionsResponses = make(map[string]*executor.HistoryHotRegions, 3)
-
-func (s *mockStoreWithMultiPD) EtcdAddrs() ([]string, error) { return s.hosts, nil }
-func (s *mockStoreWithMultiPD) TLSConfig() *tls.Config       { panic("not implemented") }
-func (s *mockStoreWithMultiPD) StartGCWorker() error         { panic("not implemented") }
-func (s *mockStoreWithMultiPD) Name() string                 { return "mockStore" }
-func (s *mockStoreWithMultiPD) Describe() string             { return "" }
-
-var _ = SerialSuites(&testHotRegionsHistoryTableSuite{testInfoschemaTableSuiteBase: &testInfoschemaTableSuiteBase{}})
-
-type testHotRegionsHistoryTableSuite struct {
-	*testInfoschemaTableSuiteBase
-	httpServers []*httptest.Server
-	startTime   time.Time
-}
-
-func (s *testHotRegionsHistoryTableSuite) SetUpSuite(c *C) {
-	s.testInfoschemaTableSuiteBase.SetUpSuite(c)
-	store := &mockStoreWithMultiPD{
-		s.store.(helper.Storage),
-		make([]string, 3),
-	}
-	// start 3 PD server with hotRegionsServer and store them in s.store
-	for i := 0; i < 3; i++ {
-		httpServer, mockAddr := s.setUpMockPDHTTPServer(c)
-		c.Assert(httpServer, NotNil)
-		s.httpServers = append(s.httpServers, httpServer)
-		store.hosts[i] = mockAddr
-	}
-	s.store = store
-	s.startTime = time.Now()
-}
-
-func writeResp(w http.ResponseWriter, resp interface{}) {
-	w.WriteHeader(http.StatusOK)
-	jsonResp, err := json.Marshal(resp)
-	if err != nil {
-		writeJSONError(w, http.StatusInternalServerError, "unable to marshal resp", err)
-		return
-	}
-	w.Write(jsonResp)
-}
-
-func writeJSONError(w http.ResponseWriter, code int, prefix string, err error) {
-	type errorResponse struct {
-		Error string `json:"error"`
-	}
-	w.WriteHeader(code)
-	if err != nil {
-		prefix += ": " + err.Error()
-	}
-	_ = json.NewEncoder(w).Encode(errorResponse{Error: prefix})
-}
-
-func hisHotRegionsHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Access-Control-Allow-Origin", "*")
-	data, err := io.ReadAll(r.Body)
-	if err != nil {
-		writeJSONError(w, http.StatusInternalServerError, "unable to read req", err)
-		return
-	}
-	r.Body.Close()
-	req := &executor.HistoryHotRegionsRequest{}
-	err = json.Unmarshal(data, req)
-	if err != nil {
-		writeJSONError(w, http.StatusInternalServerError, "unable to serialize req", err)
-		return
-	}
-	resp := &executor.HistoryHotRegions{}
-	for _, typ := range req.HotRegionTypes {
-		resp.HistoryHotRegion = append(resp.HistoryHotRegion, hotRegionsResponses[typ+r.Host].HistoryHotRegion...)
-	}
-	w.WriteHeader(http.StatusOK)
-	jsonResp, err := json.Marshal(resp)
-	if err != nil {
-		writeJSONError(w, http.StatusInternalServerError, "unable to marshal resp", err)
-		return
-	}
-	w.Write(jsonResp)
-}
-
-func (s *testHotRegionsHistoryTableSuite) setUpMockPDHTTPServer(c *C) (*httptest.Server, string) {
-	// mock PD http server
-	router := mux.NewRouter()
-	server := httptest.NewServer(router)
-	mockAddr := strings.TrimPrefix(server.URL, "http://")
-	// mock PD API
-	router.Handle(pdapi.Status, fn.Wrap(func() (interface{}, error) {
-		return struct {
-			Version        string `json:"version"`
-			GitHash        string `json:"git_hash"`
-			StartTimestamp int64  `json:"start_timestamp"`
-		}{
-			Version:        "4.0.0-alpha",
-			GitHash:        "mock-pd-githash",
-			StartTimestamp: s.startTime.Unix(),
-		}, nil
-	}))
-	// mock hisory hot regions response
-	router.HandleFunc(pdapi.HotHistory, hisHotRegionsHandler)
-	return server, mockAddr
-}
-
-func (s *testHotRegionsHistoryTableSuite) TearDownSuite(c *C) {
-	for _, server := range s.httpServers {
-		server.Close()
-	}
-	s.testInfoschemaTableSuiteBase.TearDownSuite(c)
-}
-
-func (s *testHotRegionsHistoryTableSuite) TestTiDBHotRegionsHistory(c *C) {
-	var unixTimeMs = func(s string) int64 {
-		t, err := time.ParseInLocation("2006-01-02 15:04:05", s, time.Local)
-		c.Assert(err, IsNil)
-		return t.UnixNano() / int64(time.Millisecond)
-	}
-	fullHotRegions := [][]string{
-		// mysql table_id = 11, table_name = TABLES_PRIV
-		{"2019-10-10 10:10:11", "MYSQL", "TABLES_PRIV", "11", "<nil>", "<nil>", "1", "1", "11111", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:12", "MYSQL", "TABLES_PRIV", "11", "<nil>", "<nil>", "2", "2", "22222", "0", "0", "WRITE", "99", "99", "99", "99"},
-		// mysql table_id = 21, table_name = STATS_META
-		{"2019-10-10 10:10:13", "MYSQL", "STATS_META", "21", "<nil>", "<nil>", "3", "3", "33333", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:14", "MYSQL", "STATS_META", "21", "<nil>", "<nil>", "4", "4", "44444", "0", "0", "WRITE", "99", "99", "99", "99"},
-		// table_id = 1313, deleted schema
-		{"2019-10-10 10:10:15", "UNKNOWN", "UNKNOWN", "1313", "UNKNOWN", "<nil>", "5", "5", "55555", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:16", "UNKNOWN", "UNKNOWN", "1313", "UNKNOWN", "<nil>", "6", "6", "66666", "0", "0", "WRITE", "99", "99", "99", "99"},
-		// mysql table_id = 11, index_id = 1, table_name = TABLES_PRIV, index_name = PRIMARY
-		{"2019-10-10 10:10:17", "MYSQL", "TABLES_PRIV", "11", "PRIMARY", "1", "1", "1", "11111", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:18", "MYSQL", "TABLES_PRIV", "11", "PRIMARY", "1", "2", "2", "22222", "0", "0", "WRITE", "99", "99", "99", "99"},
-		// mysql table_id = 21 ,index_id = 1, table_name = STATS_META, index_name = IDX_VER
-		{"2019-10-10 10:10:19", "MYSQL", "STATS_META", "21", "IDX_VER", "1", "3", "3", "33333", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:20", "MYSQL", "STATS_META", "21", "IDX_VER", "1", "4", "4", "44444", "0", "0", "WRITE", "99", "99", "99", "99"},
-		// mysql table_id = 21 ,index_id = 2, table_name = STATS_META, index_name = TBL
-		{"2019-10-10 10:10:21", "MYSQL", "STATS_META", "21", "TBL", "2", "5", "5", "55555", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:22", "MYSQL", "STATS_META", "21", "TBL", "2", "6", "6", "66666", "0", "0", "WRITE", "99", "99", "99", "99"},
-		// table_id = 1313, index_id = 1, deleted schema
-		{"2019-10-10 10:10:23", "UNKNOWN", "UNKNOWN", "1313", "UNKNOWN", "1", "7", "7", "77777", "0", "1", "READ", "99", "99", "99", "99"},
-		{"2019-10-10 10:10:24", "UNKNOWN", "UNKNOWN", "1313", "UNKNOWN", "1", "8", "8", "88888", "0", "0", "WRITE", "99", "99", "99", "99"},
-	}
-
-	mockDB := &model.DBInfo{}
-	pdResps := []map[string]*executor.HistoryHotRegions{
-		{
-			core.HotRegionTypeRead: {
-				HistoryHotRegion: []*executor.HistoryHotRegion{
-					// mysql table_id = 11, table_name = TABLES_PRIV
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:11"), RegionID: 1, StoreID: 1, PeerID: 11111, IsLearner: false,
-						IsLeader: true, HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 11}).StartKey,
-						EndKey:   helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 11}).EndKey,
-					},
-					// mysql table_id = 21, table_name = STATS_META
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:13"), RegionID: 3, StoreID: 3, PeerID: 33333, IsLearner: false,
-						IsLeader: true, HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 21}).StartKey,
-						EndKey:   helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 21}).EndKey,
-					},
-				},
-			},
-			core.HotRegionTypeWrite: {
-				HistoryHotRegion: []*executor.HistoryHotRegion{
-					// mysql table_id = 11, table_name = TABLES_PRIV
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:12"), RegionID: 2, StoreID: 2, PeerID: 22222, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 11}).StartKey,
-						EndKey:   helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 11}).EndKey,
-					},
-					// mysql table_id = 21, table_name = STATS_META
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:14"), RegionID: 4, StoreID: 4, PeerID: 44444, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 21}).StartKey,
-						EndKey:   helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 21}).EndKey,
-					},
-				},
-			},
-		},
-		{
-			core.HotRegionTypeRead: {
-				HistoryHotRegion: []*executor.HistoryHotRegion{
-					// table_id = 1313, deleted schema
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:15"), RegionID: 5, StoreID: 5, PeerID: 55555, IsLearner: false,
-						IsLeader: true, HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 1313}).StartKey,
-						EndKey:   helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 1313}).EndKey,
-					},
-					// mysql table_id = 11, index_id = 1, table_name = TABLES_PRIV, index_name = PRIMARY
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:17"), RegionID: 1, StoreID: 1, PeerID: 11111, IsLearner: false,
-						IsLeader: true, HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 11}, &model.IndexInfo{ID: 1}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 11}, &model.IndexInfo{ID: 1}).EndKey,
-					},
-				},
-			},
-			core.HotRegionTypeWrite: {
-				HistoryHotRegion: []*executor.HistoryHotRegion{
-					// table_id = 1313, deleted schema
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:16"), RegionID: 6, StoreID: 6, PeerID: 66666, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 1313}).StartKey,
-						EndKey:   helper.NewTableWithKeyRange(mockDB, &model.TableInfo{ID: 1313}).EndKey,
-					},
-					// mysql table_id = 11, index_id = 1, table_name = TABLES_PRIV, index_name = PRIMARY
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:18"), RegionID: 2, StoreID: 2, PeerID: 22222, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 11}, &model.IndexInfo{ID: 1}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 11}, &model.IndexInfo{ID: 1}).EndKey,
-					},
-				},
-			},
-		},
-		{
-			core.HotRegionTypeRead: {
-				HistoryHotRegion: []*executor.HistoryHotRegion{
-					// mysql table_id = 21 ,index_id = 1, table_name = STATS_META, index_name = IDX_VER
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:19"), RegionID: 3, StoreID: 3, PeerID: 33333, IsLearner: false,
-						IsLeader: true, HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 1}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 1}).EndKey,
-					},
-					// mysql table_id = 21 ,index_id = 2, table_name = STATS_META, index_name = TBL
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:21"), RegionID: 5, StoreID: 5, PeerID: 55555, IsLearner: false,
-						IsLeader: true, HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 2}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 2}).EndKey,
-					},
-					//      table_id = 1313, index_id = 1, deleted schema
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:23"), RegionID: 7, StoreID: 7, PeerID: 77777, IsLeader: true,
-						HotRegionType: "READ", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 1313}, &model.IndexInfo{ID: 1}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 1313}, &model.IndexInfo{ID: 1}).EndKey,
-					},
-				},
-			},
-			core.HotRegionTypeWrite: {
-				HistoryHotRegion: []*executor.HistoryHotRegion{
-					// mysql table_id = 21 ,index_id = 1, table_name = STATS_META, index_name = IDX_VER
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:20"), RegionID: 4, StoreID: 4, PeerID: 44444, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 1}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 1}).EndKey,
-					},
-					// mysql table_id = 21 ,index_id = 2, table_name = STATS_META, index_name = TBL
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:22"), RegionID: 6, StoreID: 6, PeerID: 66666, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 2}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 21}, &model.IndexInfo{ID: 2}).EndKey,
-					},
-					//      table_id = 1313, index_id = 1, deleted schema
-					{UpdateTime: unixTimeMs("2019-10-10 10:10:24"), RegionID: 8, StoreID: 8, PeerID: 88888, IsLearner: false,
-						IsLeader: false, HotRegionType: "WRITE", HotDegree: 99, FlowBytes: 99, KeyRate: 99, QueryRate: 99,
-						StartKey: helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 1313}, &model.IndexInfo{ID: 1}).StartKey,
-						EndKey:   helper.NewIndexWithKeyRange(mockDB, &model.TableInfo{ID: 1313}, &model.IndexInfo{ID: 1}).EndKey,
-					},
-				},
-			},
-		},
-	}
-
-	var cases = []struct {
-		conditions []string
-		reqCount   int32
-		expected   [][]string
-	}{
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-			}, // time filtered by PD, assume response suit time range, and ignore deleted schemas
-			expected: [][]string{
-				fullHotRegions[0], fullHotRegions[1], fullHotRegions[2],
-				fullHotRegions[3],
-				fullHotRegions[6], fullHotRegions[7], fullHotRegions[8],
-				fullHotRegions[9], fullHotRegions[10], fullHotRegions[11],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>=TIMESTAMP('2019-10-10 10:10:10')",
-				"update_time<=TIMESTAMP('2019-10-11 10:10:10')",
-			}, // test support of timestamp
-			expected: [][]string{
-				fullHotRegions[0], fullHotRegions[1], fullHotRegions[2],
-				fullHotRegions[3],
-				fullHotRegions[6], fullHotRegions[7], fullHotRegions[8],
-				fullHotRegions[9], fullHotRegions[10], fullHotRegions[11],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_id=11",
-			},
-			expected: [][]string{
-				fullHotRegions[0], fullHotRegions[1], fullHotRegions[6], fullHotRegions[7],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_name='TABLES_PRIV'",
-			},
-			expected: [][]string{
-				fullHotRegions[0], fullHotRegions[1], fullHotRegions[6], fullHotRegions[7],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_id=21",
-				"index_id=1",
-			},
-			expected: [][]string{
-				fullHotRegions[8], fullHotRegions[9],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_id=21",
-				"index_id=1",
-				"table_name='TABLES_PRIV'",
-			}, // table_id != table_name -> nil
-			expected: [][]string{},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_id=21",
-				"index_id=1",
-				"table_name='STATS_META'",
-			}, // table_id = table_name
-			expected: [][]string{
-				fullHotRegions[8], fullHotRegions[9],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_id=21",
-				"index_id=1",
-				"index_name='UNKNOWN'",
-			}, // index_id != index_name -> nil
-			expected: [][]string{},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"table_id=21",
-				"index_id=1",
-				"index_name='IDX_VER'",
-			}, // index_id = index_name
-			expected: [][]string{
-				fullHotRegions[8], fullHotRegions[9],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"index_id=1",
-				"index_name='IDX_VER'",
-				"table_id>=21", // unpushed down predicates 21>=21
-			},
-			expected: [][]string{
-				fullHotRegions[8], fullHotRegions[9],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"index_id=1",
-				"index_name='IDX_VER'",
-				"table_id>21", // unpushed down predicates
-			}, // 21!>21 -> nil
-			expected: [][]string{},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"index_id=1",
-				"index_name='IDX_VER'",
-				"table_id>=21", // unpushed down predicates
-				"db_name='MYSQL'",
-			},
-			expected: [][]string{
-				fullHotRegions[8], fullHotRegions[9],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"index_id=1",
-				"index_name='IDX_VER'",
-				"table_id>=21", // unpushed down predicates
-				"db_name='MYSQL'",
-				"peer_id>=33334",
-			},
-			expected: [][]string{
-				fullHotRegions[9],
-			},
-		},
-		{
-			conditions: []string{
-				"update_time>='2019-10-10 10:10:10'",
-				"update_time<='2019-10-11 10:10:10'",
-				"index_id=1",
-				"index_name='IDX_VER'",
-				"table_id>=21", // unpushed down predicates
-				"db_name='UNKNOWN'",
-			},
-			expected: [][]string{},
-		},
-	}
-
-	// mock http resp
-	store := s.store.(*mockStoreWithMultiPD)
-	for i, resp := range pdResps {
-		for k, v := range resp {
-			hotRegionsResponses[k+store.hosts[i]] = v
-		}
-	}
-	tk := testkit.NewTestKit(c, s.store)
-	for _, cas := range cases {
-		sql := "select * from information_schema.tidb_hot_regions_history"
-		if len(cas.conditions) > 0 {
-			sql = fmt.Sprintf("%s where %s", sql, strings.Join(cas.conditions, " and "))
-		}
-		result := tk.MustQuery(sql)
-		warnings := tk.Se.GetSessionVars().StmtCtx.GetWarnings()
-		c.Assert(len(warnings), Equals, 0, Commentf("unexpected warnigns: %+v, sql: %s", warnings))
-		var expected []string
-		for _, row := range cas.expected {
-			expectedRow := row
-			expected = append(expected, strings.Join(expectedRow, " "))
-		}
-		result.Check(testkit.Rows(expected...))
-	}
-}
-
-func (s *testHotRegionsHistoryTableSuite) TestTiDBHotRegionsHistoryError(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-
-	// Test without start time error
-	rs, err := tk.Exec("select * from information_schema.tidb_hot_regions_history")
-	c.Assert(err, IsNil)
-	_, err = session.ResultSetToStringSlice(context.Background(), tk.Se, rs)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "denied to scan hot regions, please specified the start time, such as `update_time > '2020-01-01 00:00:00'`")
-	c.Assert(rs.Close(), IsNil)
-
-	// Test without end time error.
-	rs, err = tk.Exec("select * from information_schema.tidb_hot_regions_history where update_time>='2019/08/26 06:18:13.011'")
-	c.Assert(err, IsNil)
-	_, err = session.ResultSetToStringSlice(context.Background(), tk.Se, rs)
-	c.Assert(err, NotNil)
-	c.Assert(err.Error(), Equals, "denied to scan hot regions, please specified the end time, such as `update_time < '2020-01-01 00:00:00'`")
-	c.Assert(rs.Close(), IsNil)
-}
-
-var regionsInfo = map[uint64]helper.RegionInfo{
-	1: {
-		ID:     1,
-		Peers:  []helper.RegionPeer{{ID: 11, StoreID: 1, IsLearner: false}, {ID: 12, StoreID: 2, IsLearner: false}, {ID: 13, StoreID: 3, IsLearner: false}},
-		Leader: helper.RegionPeer{ID: 11, StoreID: 1, IsLearner: false},
-	},
-	2: {
-		ID:     2,
-		Peers:  []helper.RegionPeer{{ID: 21, StoreID: 1, IsLearner: false}, {ID: 22, StoreID: 2, IsLearner: false}, {ID: 23, StoreID: 3, IsLearner: false}},
-		Leader: helper.RegionPeer{ID: 22, StoreID: 2, IsLearner: false},
-	},
-	3: {
-		ID:     3,
-		Peers:  []helper.RegionPeer{{ID: 31, StoreID: 1, IsLearner: false}, {ID: 32, StoreID: 2, IsLearner: false}, {ID: 33, StoreID: 3, IsLearner: false}},
-		Leader: helper.RegionPeer{ID: 33, StoreID: 3, IsLearner: false},
-	},
-}
-
-var storeRegionsInfo = &helper.RegionsInfo{
-	Count: 3,
-	Regions: []helper.RegionInfo{
-		regionsInfo[1],
-		regionsInfo[2],
-		regionsInfo[3],
-	},
-}
-
-var storesRegionsInfo = map[uint64]*helper.RegionsInfo{
-	1: storeRegionsInfo,
-	2: storeRegionsInfo,
-	3: storeRegionsInfo,
-}
-
-var _ = SerialSuites(&testTikvRegionPeersTableSuite{testInfoschemaTableSuiteBase: &testInfoschemaTableSuiteBase{}})
-
-type testTikvRegionPeersTableSuite struct {
-	*testInfoschemaTableSuiteBase
-	httpServer *httptest.Server
-	mockAddr   string
-	startTime  time.Time
-}
-
-func (s *testTikvRegionPeersTableSuite) SetUpSuite(c *C) {
-	s.testInfoschemaTableSuiteBase.SetUpSuite(c)
-	s.httpServer, s.mockAddr = s.setUpMockPDHTTPServer()
-	s.startTime = time.Now()
-}
-
-func storesRegionsInfoHandler(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id, err := strconv.Atoi(vars["id"])
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, "unable to parse id", err)
-		return
-	}
-	writeResp(w, storesRegionsInfo[uint64(id)])
-}
-
-func regionsInfoHandler(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id, err := strconv.Atoi(vars["id"])
-	if err != nil {
-		writeJSONError(w, http.StatusBadRequest, "unable to parse id", err)
-		return
-	}
-	writeResp(w, regionsInfo[uint64(id)])
-}
-func (s *testTikvRegionPeersTableSuite) TearDownSuite(c *C) {
-	s.httpServer.Close()
-	s.testInfoschemaTableSuiteBase.TearDownSuite(c)
-}
-
-func (s *testTikvRegionPeersTableSuite) setUpMockPDHTTPServer() (*httptest.Server, string) {
-	// mock PD http server
-	router := mux.NewRouter()
-	server := httptest.NewServer(router)
-	// mock store stats stat
-	mockAddr := strings.TrimPrefix(server.URL, "http://")
-	// mock PD API
-	router.Handle(pdapi.Status, fn.Wrap(func() (interface{}, error) {
-		return struct {
-			Version        string `json:"version"`
-			GitHash        string `json:"git_hash"`
-			StartTimestamp int64  `json:"start_timestamp"`
-		}{
-			Version:        "4.0.0-alpha",
-			GitHash:        "mock-pd-githash",
-			StartTimestamp: s.startTime.Unix(),
-		}, nil
-	}))
-	// mock get regionsInfo by store id
-	router.HandleFunc(pdapi.StoreRegions+"/"+"{id}", storesRegionsInfoHandler)
-	// mock get regionInfo by region id
-	router.HandleFunc(pdapi.RegionByID+"/"+"{id}", regionsInfoHandler)
-	return server, mockAddr
-}
-
-func (s *testTikvRegionPeersTableSuite) TestTikvRegionPeers(c *C) {
-	mockAddr := s.mockAddr
-	store := &mockStore{
-		s.store.(helper.Storage),
-		mockAddr,
-	}
-
-	fullRegionPeers := [][]string{
-		{"1", "11", "1", "0", "1", "NORMAL", "<nil>"},
-		{"1", "12", "2", "0", "0", "NORMAL", "<nil>"},
-		{"1", "13", "3", "0", "0", "NORMAL", "<nil>"},
-
-		{"2", "21", "1", "0", "0", "NORMAL", "<nil>"},
-		{"2", "22", "2", "0", "1", "NORMAL", "<nil>"},
-		{"2", "23", "3", "0", "0", "NORMAL", "<nil>"},
-
-		{"3", "31", "1", "0", "0", "NORMAL", "<nil>"},
-		{"3", "32", "2", "0", "0", "NORMAL", "<nil>"},
-		{"3", "33", "3", "0", "1", "NORMAL", "<nil>"},
-	}
-
-	var cases = []struct {
-		conditions []string
-		reqCount   int32
-		expected   [][]string
-	}{
-		{
-			conditions: []string{
-				"store_id in (1,2,3)",
-				"region_id in (1,2,3)",
-			},
-			expected: fullRegionPeers,
-		},
-		{
-			conditions: []string{
-				"store_id in (1,2)",
-				"region_id=1",
-			},
-			expected: [][]string{
-				fullRegionPeers[0], fullRegionPeers[1],
-			},
-		},
-		{
-			conditions: []string{
-				"store_id in (1,2)",
-				"region_id=1",
-				"is_leader=1",
-			},
-			expected: [][]string{
-				fullRegionPeers[0],
-			},
-		},
-		{
-			conditions: []string{
-				"store_id in (1,2)",
-				"region_id=1",
-				"is_leader=0",
-			},
-			expected: [][]string{
-				fullRegionPeers[1],
-			},
-		},
-		{
-			conditions: []string{
-				"store_id =1",
-				"region_id =1",
-				"is_leader =0",
-			},
-			expected: [][]string{},
-		},
-	}
-
-	tk := testkit.NewTestKit(c, store)
-	for _, cas := range cases {
-		sql := "select * from information_schema.tikv_region_peers"
-		if len(cas.conditions) > 0 {
-			sql = fmt.Sprintf("%s where %s", sql, strings.Join(cas.conditions, " and "))
-		}
-		result := tk.MustQuery(sql)
-		warnings := tk.Se.GetSessionVars().StmtCtx.GetWarnings()
-		c.Assert(len(warnings), Equals, 0, Commentf("unexpected warnigns: %+v", warnings))
-		var expected []string
-		for _, row := range cas.expected {
-			expectedRow := row
-			expected = append(expected, strings.Join(expectedRow, " "))
-		}
-		result.Check(testkit.Rows(expected...))
-	}
 }
